requirements.txt
# Once your venv is active, run:
#   pip install -r requirements.txt
opencv-python
mediapipe
pyautogui
pyyaml
